worker/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
async-timeout==4.0.3
    # via redis
billiard==4.1.0
    # via celery
celery==5.3.1
    # via -r requirements.in
click==8.1.6
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
kombu==5.3.1
    # via celery
packaging==23.1
    # via pytesseract
pillow==10.0.0
    # via
    #   -r requirements.in
    #   pytesseract
prompt-toolkit==3.0.39
    # via click-repl
pytesseract==0.3.10
    # via -r requirements.in
python-dateutil==2.8.2
    # via celery
redis==4.6.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
tzdata==2023.3
    # via celery
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchdog==3.0.0
    # via -r requirements.in
wcwidth==0.2.6
    # via prompt-toolkit
